refactor: consolidate GitHub token and release channel settings into dedicated components
