+++ src/ui/src/components/settings/release-channels/types.ts
@@ -1,0 +1,13 @@
+export interface ProcessedRelease {
+  prNumber: string;
+  tag: string;
+  version: string;
+  publishedAt: string;
+}
+
+type ReleaseChannelType = "latest" | "pr";
+
+export interface ReleaseChannel {
+  type: ReleaseChannelType;
+  channel?: string;
+}
